Refactor base page structure: move BasePage to com.base package and update imports across the project
