+++ Selenium/demo/src/main/java/com/saucedemo/pages/ProductsPage.java
@@ -2,6 +2,8 @@ package com.saucedemo.pages;
 
 import org.openqa.selenium.By;
 
+import com.base.BasePage;
+
 public class ProductsPage extends BasePage {
 
     private By productHeader = By.xpath("//span[text()='Products']");
