+++ Selenium/demo/src/main/java/utilities/Utility.java
@@ -2,7 +2,7 @@ package utilities;
 
 import org.openqa.selenium.WebDriver;
 
-import com.saucedemo.pages.BasePage;
+import com.base.BasePage;
 
 public class Utility {
 
@@ -11,8 +11,4 @@ public class Utility {
     public static void setUtilityDriver(){
         driver = BasePage.driver;
     }
-
-    //Generate a random number
-
-    //Return String to upper case
 }
